+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 See [RELEASE](./RELEASE.md) for workflow instructions.
 
+## v1.7.2
+
+### Highlights
+
+### Features
+
+### Improvements
+
+* [#6289](https://github.com/spacemeshos/go-spacemesh/pull/6289) Added write coalescing for ballots, reducing the
+  amount of time and CPU we're spending on writing ballots.
+
 ## v1.7.0
 
 ### Upgrade information